Refactored representation of in-hand acquirables.
Implemented Transient to provide an easy means to: 1) mark child objects
as the "transient" representation of an object (the object to render
while "in hand," etc.); 2) save the transform state of an object before
repositioning it; 3) restore the previous transform state.

Removed transient mesh and transient materials from acquirable in favor
of using a Transient instance on the mesh object.